Use symbol/word boundaries in match-all
This changes the way evil-multiedit-match-all works by adjusting
evil-multiedit--match-bounds (now evil-multiedit--get-occurrence) to
return a regexp as well as the match bounds. Word and symbol boundaries
are added to the regexp by default to get more consistent matching from
match-all.

This also introduces the evil-multiedit-use-symbols option which
replaces the generic thing-at-point-fn option in get-occurrences and
uses evil-word-symbol-search as a default. More work will probably need
to be done to make the use of this option apply consistently across the
package.

Fixes #10