Fix evil-multiedit--get-occurrence to properly match words
Previously, when calling a function like `evil-multiedit-match-and-next`,
multiedit would read the word at the current point but then would try to match
symbols that match that word. This presents a problem when symbols can contain
multiple words (as in the default emacs-lisp syntax table).

For example, in the following snippet...

```elisp
(setq foo-bar 123
      foo-baz 456)
```

...calling `evil-multiedit-match-and-next` while the cursor is on the first
instance of `foo` would fail as there are zero symbols matching `foo`.